+++ FakerCore/Generators/SByteGenerator.cs
@@ -4,7 +4,7 @@ public class SByteGenerator : IGenerator
 {
     public object Generate(Type type, GeneratorContext context)
     {
-        return context.Random.Next(sbyte.MinValue, sbyte.MaxValue);
+        return (sbyte)context.Random.Next(sbyte.MinValue, sbyte.MaxValue);
     }
 
     public bool CanGenerate(Type type)
