+++ FakerCore/Generators/ShortGenerator.cs
@@ -4,7 +4,7 @@ public class ShortGenerator : IGenerator
 {
     public object Generate(Type type, GeneratorContext context)
     {
-        return context.Random.Next(short.MinValue, short.MaxValue);
+        return (short)context.Random.Next(short.MinValue, short.MaxValue);
     }
 
     public bool CanGenerate(Type type)
