+++ FakerCore/Generators/UShortGenerator.cs
@@ -4,7 +4,7 @@ public class UShortGenerator : IGenerator
 {
     public object Generate(Type type, GeneratorContext context)
     {
-        return context.Random.Next(ushort.MinValue, ushort.MaxValue);
+        return (ushort)context.Random.Next(ushort.MinValue, ushort.MaxValue);
     }
 
     public bool CanGenerate(Type type)
